Refactor : dnd 관련 로직 분리
