+++ src/app/(content-layout)/[groupId]/tasklist/_tasklist/hooks/use-dnd-kit.ts
@@ -1,0 +1,42 @@
+import { DragEndEvent, MouseSensor, TouchSensor, useSensor, useSensors } from '@dnd-kit/core';
+import { arrayMove } from '@dnd-kit/sortable';
+import { TaskList, Task } from '../types/task-type';
+import { useTaskActions } from './use-task-actions';
+
+export default function useDndKit(
+  currentTasks: Task[],
+  currentTaskList: TaskList,
+  sortCurrentTasks: (orderedCurrentTasks: Task[]) => void
+) {
+  const { saveNewTaskOrder } = useTaskActions();
+
+  const sensors = useSensors(
+    useSensor(MouseSensor, {
+      activationConstraint: {
+        distance: 10,
+      },
+    }),
+    useSensor(TouchSensor, {
+      activationConstraint: {
+        delay: 250,
+        tolerance: 5,
+      },
+    })
+  );
+
+  const handleDragEnd = async (event: DragEndEvent) => {
+    const { active, over } = event;
+    if (active.id !== over?.id) {
+      const oldIndex = currentTasks.findIndex((task) => task.id === active.id);
+      const newIndex = currentTasks.findIndex((task) => task.id === over?.id);
+      if (oldIndex !== -1 && newIndex !== -1) {
+        sortCurrentTasks(arrayMove(currentTasks, oldIndex, newIndex));
+        await saveNewTaskOrder(currentTaskList!.id, Number(active.id), newIndex);
+      }
+    }
+  };
+  return {
+    sensors,
+    handleDragEnd,
+  };
+}
